requirements.txt
supabase>=1.0.0
python-dotenv>=0.19.0
numpy>=1.20.0
requests>=2.25.0
torch>=2.0.0
torchvision>=0.15.0
ftfy>=6.1.1
regex>=2023.0.0
Pillow>=9.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0.0